feat: fix logic to display welcome message username with token presence
I realized that since before it was data that was passed as props, if a user closes their tab and returns while
still logged in, the token would still exist. The user would still be logged in, but the prop of the username was not there.
This would mess up the api fetch that requires the username to be present. So, i implemented a useeffect that occurs when a token is present
to fetch the username. I also changed some backend logic to make this work as well.