feat: add a page where users will be able to roll and get a new item
I added the component to App.tsx, and added a new api that is a post request to get the new item.
I also added a function, where if the token is expired, the home page redirects to the login page, but this should change cuz its clunky.
The gacha page/component is really ugly, but css will make it really nice.