feat: change lvl to exp in db; add inventory model/controller/endpoint
I had to change the interface in front end because I am changing the db schema to have exp displayed instead of level.
After that I realized that prisma migrate dev does not work for docker, so i changed it to prisma migrate dev in the docker compose file.
Then I made some small changes to account model and to return item id as well, so i can use that in the item controller to fetch 1 specific item.