feat: add feature that enhances substats when user reaches breakpoints
The breakpoints are the remaining exp that is a state.
Once something is submitted, it checks if the input exp is the same as the remaining exp, and if it is
it runs the api fetch that increases a substat.
Next I have to make sure users see the substat change visually and make sure the shown substats are the same order.