feat: add endpoints and edit schema to have date for daily free Currency
I added endpoints to the backend and added a new column in Accounts. This is to keep the timestamp
of the last time user opened a free currency box. This is so users will be able to log in and open a free currency amount
after a certain time period. I also added a new page in the front end, which will have these currency boxes.