feat: get remove items to work; also refactor to show correct div spot
I got the remove/unequip function to work. But in order to do that, I had to make sure that the state
had some placeholders. Equipping an item that is 3rd in the array would make it show up first; so I had to ensure
that the item would be swapped for the correct spot in the array.