feat: add search functionality that links to that characters profile
Changed the logic so if the character is not found, return an empty array that will show a div that says
character not found; no more setting it as a string that says None. Added mapping to the results so pressing the name
will link to a new page. Get started on the new page. Add some css.