feat: add ranking spots on public profile; fix backend endpoint
I added a div that shows the rankings of the specific searched character. So now we can see
what ranking they are in all the different categories. I also changed the endpoint that gets
the total substats rank for a specific character since it had a body in the controller.