feat: add feature to see if item is equipped on a different character
I made a new endpoint to fetch the charactername given an id, but turned out i didnt need it.
I just added it to the character and inventory models (returning character Id and charactername string).
Now Users will be able to see when they go to equip item if it is equipped on a different character.