feat: add endpoints for account currency rankings; fix index.html
I added endpoints to fetch top 10 currency rankings for those that are above 1000,
and I added a protected endpoint to see your own individual ranking for currency for those
that are not in the top 10. I also edited index.html to make sure that the favicon would be updated
without have to manually search the url.