feat: complete exp EnhanceForm increase
I made a function that will feed the necessary exp required to the next level automatically.
This is to make the enhance system more streamlined so u can only enhance when u have enough currency
to level it up. By removing the range input, I make it easier and cleanr to level up an item.