feat: change currency increase to one button;
I changed it so you only need to input currency and then press a button to level up an item.
I realized that if i want to set it multiple levels, I need to refactor the logic that goes behind the substat increase.
For now it is simpler than what I have had before, no need for like a exp scroll bar or whatever.